backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
alembic
psycopg2-binary
pydantic
pydantic-settings
python-multipart
pdfplumber
reportlab
python-dateutil
httpx
pytest
asyncpg
authlib
python-jose[cryptography]
bcrypt
